+++ sdk/cosmos/azure-cosmos/src/test/java/com/azure/cosmos/CosmosDiagnosticsTest.java
@@ -492,7 +492,6 @@ public class CosmosDiagnosticsTest extends TestSuiteBase {
 
         assertThat(diagnostics.getClientSideRequestStatisticsList().size()).isEqualTo(itemIdList.size());
         assertThat(diagnostics.getQueryMetricsMap().values().iterator().next().getRetrievedDocumentCount()).isEqualTo(itemIdList.size());
-        assertThat(diagnostics.getFeedResponseCreationLatency()).isNotNull();
         assertThat(response.getCosmosDiagnostics().getDuration()).isNotNull();
         assertThat(response.getCosmosDiagnostics().getDuration()).isGreaterThan(Duration.ZERO);
     }